refactor: dto 디렉토리 개선. 회원가입 요청 검증 추가
