+++ src/main/java/swjungle/springboard/dto/auth/SignupResponseDto.java
@@ -1,4 +1,4 @@
-package swjungle.springboard.dto;
+package swjungle.springboard.dto.auth;
 
 public record SignupResponseDto(String msg) {
